fix: ensure float dtype in aggregate_regression_metrics for np.isfinite compatibility
- Added explicit .astype(float) after pd.to_numeric() for r2, slope, and intercept values
- Fixes TypeError when np.isfinite encounters object dtype arrays
- Ensures proper numeric filtering in aggregation step

Tested with:
- heavy_1st_expanded_D0_op_AHS_3+_1+_2+_combined_w0.csv (3 groups)
- skyline raw data_heavy_di_expanded_cleaned_AHS_3+only_rmHL0.csv (4 groups)

+++ proteomics_cli/process_prm_data.py
@@ -178,13 +178,13 @@ def aggregate_regression_metrics(grouped_df: pd.DataFrame) -> pd.Series:
     Computes mean, std, CV, and Q-test for R2, slope, and intercept.
     """
     # Convert to numeric and filter
-    r2_vals = pd.to_numeric(grouped_df['r2'], errors='coerce').dropna().values
+    r2_vals = pd.to_numeric(grouped_df['r2'], errors='coerce').dropna().astype(float).values
     r2_vals = r2_vals[np.isfinite(r2_vals)]
     
-    slope_vals = pd.to_numeric(grouped_df['slope'], errors='coerce').dropna().values
+    slope_vals = pd.to_numeric(grouped_df['slope'], errors='coerce').dropna().astype(float).values
     slope_vals = slope_vals[np.isfinite(slope_vals)]
     
-    intercept_vals = pd.to_numeric(grouped_df['intercept'], errors='coerce').dropna().values
+    intercept_vals = pd.to_numeric(grouped_df['intercept'], errors='coerce').dropna().astype(float).values
     intercept_vals = intercept_vals[np.isfinite(intercept_vals)]
     
     def safe_stats(vals):
